fix: type

+++ crates/api/src/consts.rs
@@ -1,2 +1,2 @@
 pub const SYSTEM_PROMPT_MAX_LEN: usize = 64 * 1024;
-pub const PAGE_SIZE_MAX: u32 = 100;
+pub const PAGE_SIZE_MAX: i64 = 100;
